PyWebApi/pyproject.toml
[tool.poetry]
name = "webapi"
version = "0.1.0"
description = ""
authors = ["'laixhe' <'laixhe@laixhe.com'>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
uvicorn = "^0.30.1"
fastapi = "^0.111.0"
pydantic = "^2.7.4"
pydantic-settings = "^2.3.3"
python-dotenv = "^1.0.1"
sqlalchemy = "^2.0.31"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
